requirements.txt
flask
Gitpython